피드백 반영_event 옮김, 불필요한 import 제거, viewLifecyclerOwner 안 된 부분 적용
