+++ app/src/main/java/kids/baba/mobile/presentation/view/fragment/AddGroupFragment.kt
@@ -1,14 +1,12 @@
 package kids.baba.mobile.presentation.view.fragment
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.annotation.StringRes
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.lifecycleScope
 import com.google.android.material.snackbar.Snackbar
 import dagger.hilt.android.AndroidEntryPoint
 import kids.baba.mobile.databinding.FragmentAddGroupBinding
@@ -18,8 +16,6 @@ import kids.baba.mobile.presentation.extension.repeatOnStarted
 import kids.baba.mobile.presentation.model.ColorModel
 import kids.baba.mobile.presentation.model.ColorUiModel
 import kids.baba.mobile.presentation.viewmodel.AddGroupViewModel
-import kids.baba.mobile.presentation.viewmodel.MyPageViewModel
-import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class AddGroupFragment : Fragment() {
@@ -34,8 +30,6 @@ class AddGroupFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentAddGroupBinding.inflate(inflater, container, false)
-        binding.viewmodel = viewModel
-        binding.lifecycleOwner = viewLifecycleOwner
         return binding.root
     }
 
@@ -46,8 +40,15 @@ class AddGroupFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        bindViewModel()
+
         setColorButton()
         collectEvent()
+    }
+
+    private fun bindViewModel() {
+        binding.viewmodel = viewModel
+        binding.lifecycleOwner = viewLifecycleOwner
     }
 
     private fun setColorButton() {
@@ -58,7 +59,6 @@ class AddGroupFragment : Fragment() {
                     .map { ColorUiModel(it.name, it.colorCode) }
             )
         }
-        Log.e("colors", colors.toString())
         val adapter = ColorAdapter { color ->
             viewModel.color.value = color.value
         }
