+++ app/src/main/java/kids/baba/mobile/presentation/viewmodel/InputBabiesInfoViewModel.kt
@@ -24,7 +24,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
-import java.util.StringJoiner
+import java.util.*
 import javax.inject.Inject
 
 @HiltViewModel
